requirements.txt
# Core dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
sendgrid==6.11.0
google-cloud-logging==3.11.3
google-cloud-bigquery==3.27.0
python-dotenv==1.0.1

# Testing dependencies
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-asyncio==0.24.0
httpx==0.28.1
